Demonstrate a TDD idiom for producing superb software: added a simple looping test, which failed. Did not rush to fix the problem, but added a few more tests to pinpoint the problem. In the course of doing so, discovered another defect (negative numbers). Test cases were generated for this defect as well.
Only then I proceeded to fix the code, taking care to fix one test at a
time. Of course, if a test pinpointed the problem of anotheer, braoer
test, then fixing the pinpointing test is likely to fix the broader
test.

This is OK. It is better to pinpoint a problem with an additional test,
then patching the code. The cost of having more than two tests checking
the the same code change is no cost at all.


Observe two more idioms here:

First, the use of very short, in this example, one letted identifiers.
In a context of of up to (say) seven lines of code, this technique 
is great for reducing clutter. Consdier function isPrimeCore: even if
you haven't guessed that 'c' is short for candidate, and 'd' is short
for 'divisor', the minimal short and focussed context in which these
identifiers prevail makes it easy to understand their purpose and
meaning.

Observe also the use of ternarization to hightlight the logic of
primality testing.